# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FlowForge",
    "description": "FlowForge - Low-code \u043e\u0440\u043a\u0435\u0441\u0442\u0440\u0430\u0442\u043e\u0440 \u0434\u0430\u043d\u043d\u044b\u0445 \u0441 \u0442\u0438\u043f\u0438\u0437\u0438\u0440\u043e\u0432\u0430\u043d\u043d\u044b\u043c\u0438 \u0430\u0440\u0442\u0435\u0444\u0430\u043a\u0442\u0430\u043c\u0438",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/pipelines/runs": {
      "get": {
        "tags": [
          "pipeline-runs"
        ],
        "summary": "List All Runs",
        "description": "List all pipeline runs across all pipelines.\n\nArgs:\n    limit: Max number of runs to return\n    offset: Offset for pagination\n    db: Database session\n\nReturns:\n    List of pipeline runs with count",
        "operationId": "list_all_runs_api_v1_pipelines_runs_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List All Runs Api V1 Pipelines Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines": {
      "get": {
        "tags": [
          "pipelines"
        ],
        "summary": "List Pipelines",
        "description": "List all pipelines (latest version of each).\n\nArgs:\n    db: Database session\n\nReturns:\n    List of pipelines with their latest version info",
        "operationId": "list_pipelines_api_v1_pipelines_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PipelineListItem"
                  },
                  "type": "array",
                  "title": "Response List Pipelines Api V1 Pipelines Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "pipelines"
        ],
        "summary": "Create Pipeline",
        "description": "Create a new pipeline (version 1).\n\nArgs:\n    pipeline_data: Pipeline creation data\n    db: Database session\n\nReturns:\n    Created pipeline version\n\nRaises:\n    HTTPException: If pipeline with same name exists",
        "operationId": "create_pipeline_api_v1_pipelines_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineVersionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines/{pipeline_id}": {
      "get": {
        "tags": [
          "pipelines"
        ],
        "summary": "Get Pipeline",
        "description": "Get latest version of a pipeline.\n\nArgs:\n    pipeline_id: Logical pipeline UUID\n    db: Database session\n\nReturns:\n    Current pipeline version\n\nRaises:\n    HTTPException: If pipeline not found",
        "operationId": "get_pipeline_api_v1_pipelines__pipeline_id__get",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "pipelines"
        ],
        "summary": "Update Pipeline",
        "description": "Update pipeline by creating a new version.\n\nArgs:\n    pipeline_id: Logical pipeline UUID\n    pipeline_data: Update data\n    db: Database session\n\nReturns:\n    New pipeline version\n\nRaises:\n    HTTPException: If pipeline not found or name conflict",
        "operationId": "update_pipeline_api_v1_pipelines__pipeline_id__put",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PipelineVersionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineVersionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "pipelines"
        ],
        "summary": "Delete Pipeline",
        "description": "Delete pipeline and all its versions.\n\nArgs:\n    pipeline_id: Logical pipeline UUID\n    db: Database session\n\nRaises:\n    HTTPException: If pipeline not found",
        "operationId": "delete_pipeline_api_v1_pipelines__pipeline_id__delete",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/pipelines/{pipeline_id}/run": {
      "post": {
        "tags": [
          "pipelines"
        ],
        "summary": "Run Pipeline",
        "description": "Run a pipeline asynchronously via the PipelineRunner background scheduler.\n\nCreates a PipelineRun record with status RUNNING and NodeRun records\nfor every node in the graph (status PENDING). The PipelineRunner\npolls the DB and picks up ready nodes automatically.\n\nArgs:\n    pipeline_id: Logical pipeline UUID\n    parameters: Pipeline input parameters\n    db: Database session\n\nReturns:\n    Created pipeline run (status will be RUNNING)\n\nRaises:\n    HTTPException: If pipeline not found",
        "operationId": "run_pipeline_api_v1_pipelines__pipeline_id__run_post",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "type": "object"
                  },
                  {
                    "type": "null"
                  }
                ],
                "title": "Parameters"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines/runs/{run_id}": {
      "get": {
        "tags": [
          "pipelines"
        ],
        "summary": "Get Pipeline Run",
        "description": "Get pipeline run by ID.\n\nArgs:\n    run_id: Pipeline run UUID\n    db: Database session\n\nReturns:\n    Pipeline run\n\nRaises:\n    HTTPException: If run not found",
        "operationId": "get_pipeline_run_api_v1_pipelines_runs__run_id__get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PipelineRunResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines/{pipeline_id}/runs": {
      "get": {
        "tags": [
          "pipeline-runs"
        ],
        "summary": "List Pipeline Runs",
        "description": "List all runs for a pipeline (across all versions).\n\nArgs:\n    pipeline_id: Logical pipeline UUID\n    limit: Max number of runs to return\n    offset: Offset for pagination\n    db: Database session\n\nReturns:\n    List of pipeline runs with count",
        "operationId": "list_pipeline_runs_api_v1_pipelines__pipeline_id__runs_get",
        "parameters": [
          {
            "name": "pipeline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Pipeline Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response List Pipeline Runs Api V1 Pipelines  Pipeline Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines/runs/{run_id}/detail": {
      "get": {
        "tags": [
          "pipeline-runs"
        ],
        "summary": "Get Pipeline Run Detail",
        "description": "Get a pipeline run with its node runs.\n\nArgs:\n    run_id: Pipeline run UUID\n    db: Database session\n\nReturns:\n    Pipeline run with node_runs included",
        "operationId": "get_pipeline_run_detail_api_v1_pipelines_runs__run_id__detail_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "title": "Response Get Pipeline Run Detail Api V1 Pipelines Runs  Run Id  Detail Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/pipelines/runs/{run_id}/nodes/{node_id}/logs": {
      "get": {
        "tags": [
          "pipeline-runs"
        ],
        "summary": "Get Node Run Logs",
        "description": "Get log contents for a specific node run.\n\nReads from log file on disk.\nLog path: LOG_DIR/run_logs/<version_id>/<node_id>/<run_id>.log\n\nArgs:\n    run_id: Pipeline run UUID\n    node_id: Node ID within the run\n    db: Database session\n\nReturns:\n    Log file contents",
        "operationId": "get_node_run_logs_api_v1_pipelines_runs__run_id__nodes__node_id__logs_get",
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          },
          {
            "name": "node_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Node Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Get Node Run Logs Api V1 Pipelines Runs  Run Id  Nodes  Node Id  Logs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PipelineListItem": {
        "properties": {
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "latest_version": {
            "type": "integer",
            "title": "Latest Version"
          },
          "is_current_id": {
            "type": "string",
            "title": "Is Current Id"
          },
          "node_count": {
            "type": "integer",
            "title": "Node Count",
            "default": 0
          },
          "edge_count": {
            "type": "integer",
            "title": "Edge Count",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "pipeline_id",
          "name",
          "latest_version",
          "is_current_id",
          "created_at",
          "updated_at"
        ],
        "title": "PipelineListItem",
        "description": "Summary item for pipeline list (shows latest version info)."
      },
      "PipelineRunResponse": {
        "properties": {
          "version_id": {
            "type": "string",
            "title": "Version Id"
          },
          "parameters": {
            "type": "object",
            "title": "Parameters"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "status": {
            "$ref": "#/components/schemas/RunStatus"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          }
        },
        "type": "object",
        "required": [
          "version_id",
          "id",
          "status",
          "started_at",
          "completed_at",
          "error_message"
        ],
        "title": "PipelineRunResponse",
        "description": "Schema for pipeline run response."
      },
      "PipelineVersionCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "graph_definition": {
            "type": "object",
            "title": "Graph Definition"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "PipelineVersionCreate",
        "description": "Schema for creating a new pipeline version."
      },
      "PipelineVersionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "pipeline_id": {
            "type": "string",
            "title": "Pipeline Id"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "graph_definition": {
            "type": "object",
            "title": "Graph Definition"
          },
          "is_current": {
            "type": "boolean",
            "title": "Is Current"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "pipeline_id",
          "version",
          "name",
          "description",
          "graph_definition",
          "is_current",
          "created_at",
          "updated_at"
        ],
        "title": "PipelineVersionResponse",
        "description": "Schema for pipeline version response."
      },
      "PipelineVersionUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "graph_definition": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Graph Definition"
          }
        },
        "type": "object",
        "title": "PipelineVersionUpdate",
        "description": "Schema for updating pipeline data (creates new version)."
      },
      "RunStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "success",
          "failed",
          "cancelled"
        ],
        "title": "RunStatus",
        "description": "Status of a pipeline or node run."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}